Working

+++ src/main.rs
@@ -56,7 +56,7 @@ pub struct G2Data {
     #[serde(default, deserialize_with="from_str_optional")]
     is_valid: Option<i32>,
     #[serde(default, deserialize_with="from_str_optional")]
-    ACC_X_MPS2: Option<f64>,
+    ACC_X_MPS2: Option<f32>,
     #[serde(default, deserialize_with="from_str_optional")]
     ACC_Y_MPS2: Option<f64>,
     #[serde(default, deserialize_with="from_str_optional")]
